# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "openkoutsi API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/plans/{plan_id}/workouts": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Add Workout",
        "operationId": "add_workout_api_plans__plan_id__workouts_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlannedWorkoutCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/workouts/{workout_id}": {
      "put": {
        "tags": [
          "plans"
        ],
        "summary": "Update Workout",
        "operationId": "update_workout_api_plans__plan_id__workouts__workout_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          },
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlannedWorkoutUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "plans"
        ],
        "summary": "Delete Workout",
        "operationId": "delete_workout_api_plans__plan_id__workouts__workout_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          },
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plans/{plan_id}/generate-upcoming/workouts": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Generate Upcoming Workouts",
        "description": "Synthesize structured workouts for the plan's upcoming days (no upload).\n\nFor each planned workout falling within the upcoming-week window (today\u2192+6;\nrest days and unstructured rows excluded), ensure a structured\n``WorkoutDefinition`` exists \u2014 generating one via the LLM and caching it on\n``PlannedWorkout.workout_definition_id`` when missing. The generated workouts\nshow up in the Workouts tab, where they can be reviewed, edited, and uploaded\nindividually. Returns a per-workout summary.",
        "operationId": "generate_upcoming_workouts_api_plans__plan_id__generate_upcoming_workouts_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUpcomingWorkoutsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateUpcomingWorkoutsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workouts": {
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "List workout definitions",
        "operationId": "listWorkouts",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WorkoutDefinitionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Create Workout",
        "operationId": "create_workout_api_workouts_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutDefinitionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/{workout_id}": {
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "Get Workout",
        "operationId": "get_workout_api_workouts__workout_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "workouts"
        ],
        "summary": "Update Workout",
        "operationId": "update_workout_api_workouts__workout_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutDefinitionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDefinitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workouts"
        ],
        "summary": "Delete Workout",
        "operationId": "delete_workout_api_workouts__workout_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/{workout_id}/export/{format_key}": {
      "get": {
        "tags": [
          "workouts"
        ],
        "summary": "Export Workout",
        "operationId": "export_workout_api_workouts__workout_id__export__format_key__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          },
          {
            "name": "format_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Format Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/{workout_id}/push/wahoo": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Push Workout To Wahoo",
        "description": "Push a structured workout to Wahoo as a plan + scheduled workout pair.\n\nCreates (or updates, by ``external_id``) a plan in the athlete's Wahoo\nlibrary and a workout scheduled within the today\u2192+6 day visibility window so\nit appears under \"Planned Workouts\" on ELEMNT / RIVAL devices.",
        "operationId": "push_workout_to_wahoo_api_workouts__workout_id__push_wahoo_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WahooPushRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WahooPushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateUpcomingWorkoutsRequest": {
        "properties": {
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "refresh": {
            "type": "boolean",
            "title": "Refresh",
            "default": false
          }
        },
        "type": "object",
        "title": "GenerateUpcomingWorkoutsRequest",
        "description": "Optional explicit date range (within the upcoming-week window) and refresh flag."
      },
      "GenerateUpcomingWorkoutsResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/GenerateUpcomingResultItem"
            },
            "type": "array",
            "title": "Results",
            "default": []
          }
        },
        "type": "object",
        "title": "GenerateUpcomingWorkoutsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_WorkoutDefinitionResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkoutDefinitionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[WorkoutDefinitionResponse]"
      },
...
      "PlannedWorkoutCreate": {
        "properties": {
          "week_number": {
            "type": "integer",
            "title": "Week Number"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "workout_type": {
            "type": "string",
            "title": "Workout Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "target_load": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Load"
          }
        },
        "type": "object",
        "required": [
          "week_number",
          "day_of_week",
          "workout_type"
        ],
        "title": "PlannedWorkoutCreate",
        "description": "A new planned workout added to an existing plan."
      },
      "PlannedWorkoutResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "week_number": {
            "type": "integer",
            "title": "Week Number"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "workout_type": {
            "type": "string",
            "title": "Workout Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "target_load": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Load"
          },
          "linked_activity_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Activity Ids",
            "default": []
          },
          "completed_activity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Activity Id"
          },
          "skip_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Reason"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "week_number",
          "day_of_week",
          "workout_type"
        ],
        "title": "PlannedWorkoutResponse"
      },
      "PlannedWorkoutUpdate": {
        "properties": {
          "workout_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "target_load": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Load"
          },
          "day_of_week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Week"
          },
          "week_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Number"
          }
        },
        "type": "object",
        "title": "PlannedWorkoutUpdate",
        "description": "Editable fields of a single planned workout."
      },
...
      "WahooPushRequest": {
        "properties": {
          "starts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts",
            "description": "When to schedule the workout. Must be within today\u2192+6 days. Defaults to now."
          }
        },
        "type": "object",
        "title": "WahooPushRequest"
      },
      "WahooPushResponse": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "workout_id": {
            "type": "string",
            "title": "Workout Id"
          },
          "starts": {
            "type": "string",
            "format": "date-time",
            "title": "Starts"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "workout_id",
          "starts"
        ],
        "title": "WahooPushResponse"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}